requirements.txt
# use conda install:
torch~=2.0.1
torchvision~=0.15.2
tensorboard～=2.15.0
scikit-learn # conda install -c conda-forge scikit-learn
jupyter  # conda install jupyter
seaborn～=0.13.0
pandas~=2.0.3
numpy~=1.25.2
# use pip install:
geopandas~=0.12.0
pyogrio
shapely~=2.0.1
selenium~=4.14.0
streetview~=0.0.6
pypushdeer~=0.0.3
torchmetrics~=0.9.3
tqdm
torchinfo~=1.8.0
python-dotenv
osmnx==1.6.0
fake_useragent
webdriver_manager~=4.0.1
pillow~=9.4.0

# 可选：
pymongo~=4.5.0
torchcam~=0.4.0
mlxtend~=0.22.0
# git+https://github.com/CSAILVision/semantic-segmentation-pytorch.git@master
# pysal~=2.0.0
# palettable~=3.3.0

